(PC-29416) feat(offer): extract reaction section
