+++ src/features/offer/components/OfferBody/OfferBody.tsx
@@ -1,11 +1,9 @@
 import { useNavigation } from '@react-navigation/native'
-import React, { FunctionComponent, useMemo, useState } from 'react'
+import React, { FunctionComponent, useState } from 'react'
 import { View } from 'react-native'
 import styled, { useTheme } from 'styled-components/native'
 
 import { CategoryIdEnum, OfferResponseV2 } from 'api/gen'
-import { useAuthContext } from 'features/auth/context/AuthContext'
-import { useBookings } from 'features/bookings/api'
 import { UseNavigationType } from 'features/navigation/RootNavigator/types'
 import { OfferAbout } from 'features/offer/components/OfferAbout/OfferAbout'
 import { OfferArtists } from 'features/offer/components/OfferArtists/OfferArtists'
@@ -13,7 +11,7 @@ import { OfferCTAButton } from 'features/offer/components/OfferCTAButton/OfferCT
 import { OfferMessagingApps } from 'features/offer/components/OfferMessagingApps/OfferMessagingApps'
 import { OfferPlace } from 'features/offer/components/OfferPlace/OfferPlace'
 import { OfferPrice } from 'features/offer/components/OfferPrice/OfferPrice'
-import { OfferReactions } from 'features/offer/components/OfferReactions/OfferReactions'
+import { OfferReactionSection } from 'features/offer/components/OfferReactionSection/OfferReactionSection'
 import { OfferSummaryInfoList } from 'features/offer/components/OfferSummaryInfoList/OfferSummaryInfoList'
 import { OfferTitle } from 'features/offer/components/OfferTitle/OfferTitle'
 import { OfferVenueButton } from 'features/offer/components/OfferVenueButton/OfferVenueButton'
@@ -30,11 +28,9 @@ import { getOfferPrices } from 'features/offer/helpers/getOfferPrice/getOfferPri
 import { getOfferTags } from 'features/offer/helpers/getOfferTags/getOfferTags'
 import { useArtistResults } from 'features/offer/helpers/useArtistResults/useArtistResults'
 import { useOfferSummaryInfoList } from 'features/offer/helpers/useOfferSummaryInfoList/useOfferSummaryInfoList'
-import { ReactionChoiceValidation } from 'features/reactions/components/ReactionChoiceValidation/ReactionChoiceValidation'
 import { analytics } from 'libs/analytics'
 import { useFeatureFlag } from 'libs/firebase/firestore/featureFlags/useFeatureFlag'
 import { RemoteStoreFeatureFlags } from 'libs/firebase/firestore/types'
-import { useRemoteConfigContext } from 'libs/firebase/remoteConfig/RemoteConfigProvider'
 import { Subcategory } from 'libs/subcategories/types'
 import { isNullOrUndefined } from 'shared/isNullOrUndefined/isNullOrUndefined'
 import { SurveyModal } from 'ui/components/modals/SurveyModal'
@@ -61,34 +57,9 @@ export const OfferBody: FunctionComponent<Props> = ({
   const { visible, showModal, hideModal } = useModal()
   const { navigate } = useNavigation<UseNavigationType>()
 
-  const { isLoggedIn, user } = useAuthContext()
   const [surveyModalState, setSurveyModalState] = useState(DEFAULT_SURVEY_MODAL_DATA)
   const hasFakeDoorArtist = useFeatureFlag(RemoteStoreFeatureFlags.FAKE_DOOR_ARTIST)
   const hasArtistPage = useFeatureFlag(RemoteStoreFeatureFlags.WIP_ARTIST_PAGE)
-  const isReactionFeatureActive = useFeatureFlag(RemoteStoreFeatureFlags.WIP_REACTION_FEATURE)
-  const { reactionCategories } = useRemoteConfigContext()
-  const { data: bookings } = useBookings()
-
-  const endedBookingWithoutCancellation = bookings?.ended_bookings?.find(
-    (booking) => booking.stock.offer.id === offer.id && !booking.cancellationDate
-  )
-
-  const userBooking = useMemo(
-    () =>
-      bookings?.ended_bookings?.find(
-        (booking) => booking.stock.offer.id === offer.id && !booking.cancellationDate
-      ),
-    [bookings, offer.id]
-  )
-
-  const shouldDisplayReactionButtons =
-    isLoggedIn &&
-    user?.isBeneficiary &&
-    reactionCategories.categories.includes(subcategory.nativeCategoryId) &&
-    !!endedBookingWithoutCancellation
-
-  const canDisplayReactionSection =
-    isReactionFeatureActive && (offer.reactionsCount.likes > 0 || shouldDisplayReactionButtons)
 
   const shouldDisplayFakeDoorArtist =
     hasFakeDoorArtist && FAKE_DOOR_ARTIST_SEARCH_GROUPS.includes(subcategory.searchGroupName)
@@ -160,21 +131,7 @@ export const OfferBody: FunctionComponent<Props> = ({
 
           {prices ? <OfferPrice prices={prices} /> : null}
 
-          {canDisplayReactionSection ? (
-            <ViewGap gap={4}>
-              <OfferReactions
-                user={user}
-                isLoggedIn={isLoggedIn}
-                offer={offer}
-                userCanReact={shouldDisplayReactionButtons}
-                userBooking={userBooking}
-              />
-
-              {shouldDisplayReactionButtons ? (
-                <ReactionChoiceValidation handleOnPressReactionButton={() => ''} />
-              ) : null}
-            </ViewGap>
-          ) : null}
+          <OfferReactionSection offer={offer} subcategory={subcategory} />
 
           <GroupWithSeparator
             showTopComponent={offer.venue.isPermanent}
